refactor: create AuthService.ts with login and register methods
