+++ src/pages/login/index.tsx
@@ -54,9 +54,11 @@ export default function LoginPage() {
         setFormErrorMessage(error.message);
         setFormErrorAction(error.action);
         setIsLoading(false);
+        setIsDisabled(true);
         setTimeout(() => {
           setFormErrorMessage("");
           setFormErrorAction("");
+          setIsDisabled(false);
         }, 5000);
       }
     }
