+++ src/services/AuthService.ts
@@ -1,10 +1,24 @@
-import { InternalServerError, LoginServiceError, TokenServiceError } from "@/lib/CustomErrors";
+import {
+  InternalServerError,
+  LoginServiceError,
+  RegisterServiceError,
+  TokenServiceError,
+  UserValidationsError,
+} from "@/lib/CustomErrors";
 import { IHasher } from "@/lib/Hasher";
 import { ITokenService } from "@/lib/TokenService";
+import User from "@/models/User";
 import { IAuthRepository } from "@/repository/AuthRepository";
 import { IUserRepository } from "@/repository/UserRepository";
 
-export class LoginService {
+interface INewUserData {
+  name: string;
+  email: string;
+  password: string;
+  passwordConfirmation: string;
+}
+
+export default class AuthService {
   constructor(
     private readonly userRepository: IUserRepository,
     private readonly authRepository: IAuthRepository,
@@ -12,7 +26,41 @@ export class LoginService {
     private readonly tokenService: ITokenService,
   ) {}
 
-  async login(email: string, password: string) {
+  async register({ name, email, password, passwordConfirmation }: INewUserData): Promise<void> {
+    try {
+      const user = new User(name, email, password, passwordConfirmation);
+      const newUser = user.getUser()!;
+
+      const isUserExists = await this.userRepository.findUserByEmail(email);
+      if (isUserExists) {
+        throw new RegisterServiceError("Credenciais inválidas.", "Tente Novamente.", 400, true);
+      }
+
+      const hashedPassword = await this.hasher.encrypt(password);
+
+      const userWithHashedPassword = {
+        ...newUser,
+        password: hashedPassword,
+      };
+
+      await this.userRepository.createUser(userWithHashedPassword);
+    } catch (error) {
+      if (error instanceof UserValidationsError || error instanceof RegisterServiceError) {
+        throw error;
+      }
+
+      console.error("Error during user register: ", error);
+
+      throw new InternalServerError(
+        "Ocorreu um Erro inesperado ao tentar realizar o cadastro",
+        "Tente novamente mais tarde",
+        500,
+        true,
+      );
+    }
+  }
+
+  async login(email: string, password: string): Promise<string> {
     try {
       const isUserExists = await this.userRepository.findUserByEmail(email);
       if (!isUserExists) {
